fix(harness): tighten verification and finalize dag node refactor
