+++ .design/open-tasks/README.md
@@ -1,35 +1,19 @@
-# 🗂️ Open Tasks (최신)
+# Open Tasks 가이드
 
-이 디렉토리는 **남은 작업 목록**만 관리합니다.
+이 디렉토리는 “남아 있는 일”과 “참고용 감사 산출물”만 둡니다.
 
-## 파일
-- `CURRENT-TASKS.md`: 남은 작업([ ])만 포함
-- `FUTURE-PROJECTS.md`: 장기 계획(스펙 문서 아님)
-- **1-2주 내 완료**: CURRENT-TASKS.md에 추가
-- **진행중 Remote 작업**: REMOTE-SYSTEM.md에 추가
-- **3개월+ 장기**: FUTURE-PROJECTS.md에 추가
+## 파일 역할
+- `CURRENT-TASKS.md`: 1-2주 내 처리할 실제 남은 작업만 기록
+- `FUTURE-PROJECTS.md`: 장기 구상과 참고용 테마
+- `ssot-duplicate-declarations-v*.json`: 중복/SSOT 감사 스냅샷
 
----
+## 운영 원칙
+- 완료된 작업은 `CURRENT-TASKS.md`에서 제거합니다.
+- 현재 상태 규약은 각 workspace의 `docs/SPEC.md`가 소유합니다.
+- 장기 항목이 실행 단계로 올라오면 `CURRENT-TASKS.md`로 승격합니다.
+- 감사 JSON은 판단 보조 자료이며, 스펙이나 작업 체크리스트를 대체하지 않습니다.
 
-## ✅ 통합 완료 내역
-
-### 삭제된 파일 (통합 완료)
-- ~~agent-event-normalization-checklist.md~~ → CURRENT-TASKS.md
-- ~~workflow-open-tasks.md~~ → CURRENT-TASKS.md
-- ~~OPEN-TASKS.md~~ → CURRENT-TASKS.md
-- ~~IMPLEMENTATION-CHECKLIST.md~~ → REMOTE-SYSTEM.md
-- ~~npm-deployment-checklist.md~~ → 삭제 (Phase 1-5 완료)
-- ~~planning-implementation-checklist.md~~ → FUTURE-PROJECTS.md
-- ~~ROADMAP.md~~ → FUTURE-PROJECTS.md
-
-### 통합 원칙
-1. **시간대별 분류**: 즉시(1-2주) / 진행중(Remote) / 장기(3개월+)
-2. **중복 제거**: 동일 작업 병합
-3. **우선순위 명확화**: Priority 1-4 구분
-4. **완료 작업 정리**: 각 문서 상단에 ✅ 섹션
-
----
-
-**마지막 정리**: 2025-10-16 by AI Assistant
-**통합 이유**: 파일 수 줄이기 (7개 → 3개), 작업 명확성 향상
-
+## 정리 기준
+- 단기 실행 계획은 하나의 문서(`CURRENT-TASKS.md`)로 유지합니다.
+- 완료 이력은 여기 누적하지 않습니다.
+- 임시 분석 메모는 `.design/tmp/`에 두고, 정식 합의가 끝나면 삭제하거나 owner 문서로 승격합니다.
