+++ scripts/harness/review-change.mjs
@@ -111,7 +111,7 @@ function renderReportPath(reportPath) {
 async function main() {
   const options = parseScopeArgs(process.argv.slice(2));
   const scopes = await listWorkspaceScopes();
-  const changedFiles = detectChangedFiles();
+  const changedFiles = detectChangedFiles(options.baseRef);
   const scopeFiles = mapFilesToScopes(changedFiles, scopes);
   const selectedScopes = options.scopeTokens.length > 0
     ? resolveRequestedScopes(options.scopeTokens, scopes)
@@ -195,8 +195,14 @@ async function main() {
 
     scopeRecommendations.push({
       scope,
-      includeScenarios: hasScenarioVerification && (classification.hasScenarioChanges || (classification.hasSourceChanges && hasExamplesDir)),
-      includeRecord: hasScenarioRecord && (classification.hasScenarioChanges || (classification.hasSourceChanges && hasExamplesDir)),
+      includeScenarios: hasScenarioVerification && (
+        classification.hasScenarioChanges
+        || ((classification.hasSourceChanges || classification.hasConfigChanges) && hasExamplesDir)
+      ),
+      includeRecord: hasScenarioRecord && (
+        classification.hasScenarioChanges
+        || ((classification.hasSourceChanges || classification.hasConfigChanges) && hasExamplesDir)
+      ),
     });
   }
